[BUGFIX] Exif extraction via PHP does not work
When merging cc_* extensions to svmetaextract one line of code was
deleted in sv2. Without it the EXIF meta data extraction via PHP will
not work. Problem is that the collected meta data will not be saved and
therefore not displayed in the "Additional data" tab of the DAM record.

Release: 1.1.0

Change-Id: I0a8bb84ba98bb9fba72efaff78056a68d199c9db
Fixes: #43060