[FEATURE] Provide a switch for each service
Adds EM configuration for each service.
With the new situation of having multiple metadata extraction services
in one extension that partially do the same thing with just different
libraries, it would be good to have a switch for each service in the
extension configuration because different people may have different
opinions on which service is the best.

Release: 1.1.0

Change-Id: Ie2c94e6a901560380c6d0a8075ab3db95327741e
Fixes: #43169
Reviewed-on: https://review.typo3.org/17316
Reviewed-by: Christian Weiske
Tested-by: Christian Weiske
Reviewed-by: Felix Nagel
Tested-by: Felix Nagel
Reviewed-by: Lorenz Ulrich
Tested-by: Lorenz Ulrich